// utils/interfaces.ts
export interface Character {
	homeworldName: string;
	vehicleNames: string[];
	starshipNames: string[];
	filmNames: string[];
	loading: boolean;
  speciesNames: string[];
  url: string;
  name?: string;
  height?: string;
  mass?: string;
  hair_color?: string;
  skin_color?: string;
  eye_color?: string;
  birth_year?: string;
  gender?: string;
  homeworld?: string;
  films?: string[];
  species?: string[];
  vehicles?: string[];
  starships?: string[];
  created?: string;
  edited?: string;
}